+++ test/KintoWallet.t.sol
@@ -752,7 +752,7 @@ contract KintoWalletTest is AATestScaffolding, UserOp {
         // Mint NFT to new owner and burn old
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(
             _kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint16[] memory traits = new uint16[](0);
+        uint8[] memory traits = new uint8[](0);
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         sigdata = _auxCreateSignature(_kintoIDv1, _owner, _owner, 1, block.timestamp + 1000);
@@ -808,7 +808,7 @@ contract KintoWalletTest is AATestScaffolding, UserOp {
         // Mint NFT to new owner
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(
             _kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint16[] memory traits = new uint16[](0);
+        uint8[] memory traits = new uint8[](0);
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         vm.stopPrank();
@@ -868,7 +868,7 @@ contract KintoWalletTest is AATestScaffolding, UserOp {
         // Mint NFT to new owner and burn old
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(
             _kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint16[] memory traits = new uint16[](0);
+        uint8[] memory traits = new uint8[](0);
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         sigdata = _auxCreateSignature(_kintoIDv1, _owner, _owner, 1, block.timestamp + 1000);
